Fix machine custom feature validation
