+++ modules/python/tests/test_aks_machine_client.py
@@ -148,62 +148,6 @@ class TestAKSMachineClient(unittest.TestCase):
         """Non-multiple-of-1000 counts >= 1000 stay literal."""
         self.assertEqual(AKSMachineClient._get_machine_name_prefix(1500), "scale1500")
 
-    def test_scriptless_enabled_value_from_custom_features(self):
-        """DisableSelfContainedVHD marks the run as not scriptless-enabled."""
-        self.assertEqual(AKSMachineClient._scriptless_enabled_value(None), "yes")
-        self.assertEqual(AKSMachineClient._scriptless_enabled_value(""), "yes")
-        self.assertEqual(
-            AKSMachineClient._scriptless_enabled_value("SomeOtherFeature"), "yes"
-        )
-        self.assertEqual(
-            AKSMachineClient._scriptless_enabled_value(
-                "SomeOtherFeature, DisableSelfContainedVHD"
-            ),
-            "no",
-        )
-        self.assertEqual(
-            AKSMachineClient._scriptless_enabled_value(
-                "SomeOtherFeature DisableSelfContainedVHD"
-            ),
-            "yes",
-        )
-        self.assertEqual(
-            AKSMachineClient._scriptless_enabled_value(
-                "SomeOtherFeature;DisableSelfContainedVHD"
-            ),
-            "yes",
-        )
-
-    def test_custom_feature_headers_pass_through_comma_delimited_list(self):
-        """Custom feature header uses the caller-provided comma-delimited string."""
-        self.assertEqual(AKSMachineClient._custom_feature_headers(None), {})
-        self.assertEqual(
-            AKSMachineClient._custom_feature_headers(
-                "SomeOtherFeature, DisableSelfContainedVHD"
-            ),
-            {
-                "AKSHTTPCustomFeatures": (
-                    "SomeOtherFeature, DisableSelfContainedVHD"
-                ),
-            },
-        )
-        self.assertEqual(
-            AKSMachineClient._custom_feature_headers(
-                "  SomeOtherFeature, DisableSelfContainedVHD  "
-            ),
-            {
-                "AKSHTTPCustomFeatures": (
-                    "SomeOtherFeature, DisableSelfContainedVHD"
-                ),
-            },
-        )
-        self.assertEqual(
-            AKSMachineClient._scriptless_enabled_value(
-                "SomeOtherFeature, DisableSelfContainedVHD"
-            ),
-            "no",
-        )
-
     # ---- scale_machine: non-batch path ----
 
     @mock.patch.object(AKSMachineClient, "_wait_for_machine_node_readiness")
@@ -829,297 +773,6 @@ class TestAKSMachineClient(unittest.TestCase):
             expected_names={"m1"},
         )
         self.assertGreater(self.mock_k8s.get_ready_nodes.call_count, 1)
-
-    # ---- _scale_machine_batch ----
-
-    def test_scale_machine_batch_partitions_and_aggregates(self):
-        """Names sharded across workers; all successful slices aggregated."""
-        request = SimpleNamespace(
-            agent_pool_name="apool",
-            cluster_name="fake-cluster",
-            resource_group="fake-rg",
-            vm_size="Standard_D2_v3",
-            timeout=60,
-            machine_workers=2,
-        )
-        names = ["m-1", "m-2", "m-3", "m-4"]
-        with mock.patch.object(
-            AKSMachineClient,
-            "_create_batch_machines",
-            side_effect=lambda req, chunk, worker_id: list(chunk),
-        ) as mock_create_batch:
-            successful = self.client._scale_machine_batch(request, names)
-        self.assertEqual(set(successful), set(names))
-        self.assertEqual(mock_create_batch.call_count, 2)
-        chunk_lengths = [
-            len(call.args[1]) for call in mock_create_batch.call_args_list
-        ]
-        self.assertEqual(sorted(chunk_lengths), [2, 2])
-
-    def test_scale_machine_batch_per_worker_failure_isolated(self):
-        """One worker raising does not poison the other worker's success list."""
-        request = SimpleNamespace(
-            agent_pool_name="apool",
-            cluster_name="fake-cluster",
-            resource_group="fake-rg",
-            vm_size="Standard_D2_v3",
-            timeout=60,
-            machine_workers=2,
-        )
-        names = ["m-1", "m-2", "m-3", "m-4"]
-        counter = itertools.count()
-
-        def fake_create(req, chunk, worker_id):  # pylint: disable=unused-argument
-            if next(counter) == 0:
-                raise RuntimeError("first worker boom")
-            return list(chunk)
-
-        with mock.patch.object(
-            AKSMachineClient, "_create_batch_machines", side_effect=fake_create
-        ):
-            successful = self.client._scale_machine_batch(request, names)
-        # Exactly one worker's slice survives.
-        self.assertEqual(len(successful), 2)
-
-    def test_scale_machine_batch_rejects_non_exact_multiple(self):
-        """scale_machine_count must be an exact multiple of machine_workers."""
-        request = SimpleNamespace(
-            agent_pool_name="apool",
-            cluster_name="fake-cluster",
-            resource_group="fake-rg",
-            vm_size="Standard_D2_v3",
-            timeout=60,
-            machine_workers=3,
-        )
-        names = ["m-1", "m-2", "m-3", "m-4"]  # 4 not divisible by 3
-        with self.assertRaises(ValueError):
-            self.client._scale_machine_batch(request, names)
-
-    def test_scale_machine_batch_rejects_non_positive_workers(self):
-        """machine_workers must be positive."""
-        request = SimpleNamespace(
-            agent_pool_name="apool",
-            cluster_name="fake-cluster",
-            resource_group="fake-rg",
-            vm_size="Standard_D2_v3",
-            timeout=60,
-            machine_workers=0,
-        )
-        with self.assertRaises(ValueError):
-            self.client._scale_machine_batch(request, ["m-1"])
-
-    def test_scale_machine_batch_rejects_calculated_batch_over_limit(self):
-        """Fail fast before ARM when calculated per-worker batch size exceeds 50."""
-        request = SimpleNamespace(
-            agent_pool_name="apool",
-            cluster_name="fake-cluster",
-            resource_group="fake-rg",
-            vm_size="Standard_D2_v3",
-            timeout=60,
-            machine_workers=10,
-        )
-        names = [f"m-{i}" for i in range(1, 1001)]
-        with mock.patch.object(
-            AKSMachineClient, "_create_batch_machines"
-        ) as mock_create_batch:
-            with self.assertRaisesRegex(ValueError, "calculated batch size"):
-                self.client._scale_machine_batch(request, names)
-        mock_create_batch.assert_not_called()
-
-    def test_scale_machine_batch_allows_calculated_batch_at_limit(self):
-        """A calculated per-worker batch size of exactly 50 is allowed."""
-        request = SimpleNamespace(
-            agent_pool_name="apool",
-            cluster_name="fake-cluster",
-            resource_group="fake-rg",
-            vm_size="Standard_D2_v3",
-            timeout=60,
-            machine_workers=20,
-        )
-        names = [f"m-{i}" for i in range(1, 1001)]
-        with mock.patch.object(
-            AKSMachineClient,
-            "_create_batch_machines",
-            side_effect=lambda req, chunk, worker_id: list(chunk),
-        ) as mock_create_batch:
-            successful = self.client._scale_machine_batch(request, names)
-        self.assertEqual(set(successful), set(names))
-        self.assertEqual(mock_create_batch.call_count, 20)
-        self.assertTrue(
-            all(len(call.args[1]) == 50 for call in mock_create_batch.call_args_list)
-        )
-
-    # ---- _create_batch_machines ----
-
-    def test_create_batch_machines_empty_chunk_returns_empty(self):
-        """Empty chunk -> early return, no HTTP call."""
-        request = SimpleNamespace(
-            agent_pool_name="apool",
-            cluster_name="fake-cluster",
-            resource_group="fake-rg",
-            vm_size="Standard_D2_v3",
-            timeout=60,
-        )
-        with mock.patch.object(
-            AKSMachineClient, "_make_batch_request"
-        ) as mock_make_batch:
-            result = self.client._create_batch_machines(request, [], 0)
-        self.assertEqual(result, [])
-        mock_make_batch.assert_not_called()
-
-    def test_create_batch_machines_header_shape(self):
-        """BatchPutMachine header carries vmSkus envelope + batchMachines
-        using machineName (NOT name) keys for the *additional* machines only."""
-        request = SimpleNamespace(
-            agent_pool_name="apool",
-            cluster_name="fake-cluster",
-            resource_group="fake-rg",
-            vm_size="Standard_D2_v3",
-            timeout=60,
-            aks_http_custom_features="SomeOtherFeature, DisableSelfContainedVHD",
-        )
-        with mock.patch.object(
-            AKSMachineClient, "_make_batch_request"
-        ) as mock_make_batch:
-            result = self.client._create_batch_machines(
-                request, ["m-1", "m-2", "m-3"], chunk_idx=0
-            )
-        self.assertEqual(result, ["m-1", "m-2", "m-3"])
-        mock_make_batch.assert_called_once()
-        # Inspect the batch_header_value kwarg.
-        import json as _json  # pylint: disable=import-outside-toplevel
-        kwargs = mock_make_batch.call_args.kwargs
-        self.assertEqual(
-            kwargs["aks_http_custom_features"],
-            "SomeOtherFeature, DisableSelfContainedVHD",
-        )
-        header_value = kwargs["batch_header_value"]
-        parsed = _json.loads(header_value)
-        # vmSkus is wrapped in {"value": [...]}
-        self.assertIn("vmSkus", parsed)
-        self.assertIn("value", parsed["vmSkus"])
-        self.assertEqual(len(parsed["vmSkus"]["value"]), 1)
-        self.assertEqual(
-            parsed["vmSkus"]["value"][0]["name"], "Standard_D2_v3"
-        )
-        # First machine of chunk is created from the URL+body; only the rest
-        # appear in batchMachines.
-        self.assertEqual(
-            parsed["batchMachines"],
-            [{"machineName": "m-2"}, {"machineName": "m-3"}],
-        )
-
-    def test_create_batch_machines_rejects_oversized_chunk(self):
-        """BatchPutMachine rejects more than 50 machines per request client-side."""
-        request = SimpleNamespace(
-            agent_pool_name="apool",
-            cluster_name="fake-cluster",
-            resource_group="fake-rg",
-            vm_size="Standard_D2_v3",
-            timeout=60,
-        )
-        names = [f"m-{i}" for i in range(1, 52)]
-        with mock.patch.object(
-            AKSMachineClient, "_make_batch_request"
-        ) as mock_make_batch:
-            with self.assertRaises(ValueError):
-                self.client._create_batch_machines(request, names, chunk_idx=0)
-        mock_make_batch.assert_not_called()
-
-    # ---- _make_batch_request ----
-
-    def test_make_batch_request_2xx_returns(self):
-        """2xx response returns without raising; one HTTP call made."""
-        with mock.patch.object(
-            self.client._session, "request"
-        ) as mock_request:
-            mock_request.return_value.status_code = 200
-            self.client._make_batch_request(
-                "PUT",
-                "https://fake/url",
-                {"k": "v"},
-                timeout=30,
-                batch_header_value="{}",
-            )
-        mock_request.assert_called_once()
-
-    def test_make_batch_request_sends_custom_feature_header(self):
-        """BatchPutMachine PUTs include AKSHTTPCustomFeatures when configured."""
-        with mock.patch.object(
-            self.client._session, "request"
-        ) as mock_request:
-            mock_request.return_value.status_code = 202
-            self.client._make_batch_request(
-                "PUT",
-                "https://fake/url",
-                {"k": "v"},
-                timeout=30,
-                batch_header_value="{}",
-                aks_http_custom_features="SomeOtherFeature, DisableSelfContainedVHD",
-            )
-        headers = mock_request.call_args.kwargs["headers"]
-        self.assertEqual(headers["BatchPutMachine"], "{}")
-        self.assertEqual(
-            headers["AKSHTTPCustomFeatures"],
-            "SomeOtherFeature, DisableSelfContainedVHD",
-        )
-
-    def test_make_batch_request_non_2xx_non_429_raises(self):
-        """500 -> RuntimeError, no retry."""
-        with mock.patch.object(
-            self.client._session, "request"
-        ) as mock_request:
-            mock_request.return_value.status_code = 500
-            mock_request.return_value.text = "boom"
-            with self.assertRaises(RuntimeError):
-                self.client._make_batch_request(
-                    "PUT",
-                    "https://fake/url",
-                    {},
-                    timeout=30,
-                    batch_header_value="{}",
-                )
-        self.assertEqual(mock_request.call_count, 1)
-
-    def test_make_batch_request_429_then_success(self):
-        """429 then 2xx (201) -> succeeds after 1 retry."""
-        responses = [
-            mock.MagicMock(status_code=429),
-            mock.MagicMock(status_code=201),
-        ]
-        with mock.patch.object(
-            self.client._session, "request", side_effect=responses,
-        ) as mock_request, mock.patch(
-            "clients.aks_machine_client.time.sleep"
-        ):
-            self.client._make_batch_request(
-                "PUT",
-                "https://fake/url",
-                {},
-                timeout=30,
-                batch_header_value="{}",
-            )
-        self.assertEqual(mock_request.call_count, 2)
-
-    def test_make_batch_request_429_exhausted_raises(self):
-        """429 across the full retry budget -> RuntimeError."""
-        with mock.patch.object(
-            self.client._session, "request"
-        ) as mock_request, mock.patch(
-            "clients.aks_machine_client.time.sleep"
-        ):
-            mock_request.return_value.status_code = 429
-            mock_request.return_value.text = "rate limited"
-            with self.assertRaises(RuntimeError):
-                self.client._make_batch_request(
-                    "PUT",
-                    "https://fake/url",
-                    {},
-                    timeout=30,
-                    batch_header_value="{}",
-                )
-        # _BATCH_429_MAX_RETRIES == 5 total attempts (no extra initial call)
-        self.assertEqual(mock_request.call_count, 5)
 
     # ---- Session connection pool ----
